Definindo o registro dos logs na applications
