+++ src/routes/logs.js
@@ -1,14 +1,14 @@
 const express = require('express')
 const router = express.Router()
 const controller = require('../controllers/logs')
-const userToken = require('../middlewares/userToken')
+const appToken = require('../middlewares/appToken')
 const queryParams = require('../middlewares/queryParam')
 
 router.get('/', queryParams.validate, controller.getAll)
 
 router.get('/:logId', controller.getById)
 
-router.post('/', userToken.validate, controller.create)
+router.post('/', appToken.validate, controller.create)
 
 router.delete('/:logId', controller.delete)
